+++ CHANGELOG.md
@@ -8,6 +8,17 @@ Upgrade version:
 Upgrade library dependancies (if required):
 - python3 -m pip install -r requirements.txt -U
 
+## [3.2.9] - 2021-08-16
+
+### Added
+
+-- 'nosellminpcnt' to specify minimum margin to not sell
+-- 'nosellmaxpcnt' to specify maximum margin to not sell
+
+### Changed
+
+-- Various small bug fixes
+
 ## [3.2.8] - 2021-08-12
 
 ### Changed